calypso_AgentiArena/whale-agent/requirements.txt
fastapi
uvicorn
langgraph
pydantic
google-generativeai
requests
